Safari-609.1.11.1/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-allowall.html

<!DOCTYPE html>
<html>
<head>
    <script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
        testRunner.dumpResourceLoadCallbacks();
    }
    </script>
</head>
<body>
    <p>The frame below should load, and 'ALLOWALL' should be accepted as a valid header.</p>
    <iframe style="width:500px; height:500px" src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-allowall.cgi"></iframe>
</body>
</html>
