LayoutTests/platform/chromium/http/tests/security/XFrameOptions/x-frame-options-allowall-expected.txt

http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-allowall.cgi - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-allowall.cgi, main document URL http://127.0.0.1:8000/security/XFrameOptions/x-frame-options-allowall.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-allowall.cgi - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-allowall.cgi, http status code 200>
http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-allowall.cgi - didFinishLoading
The frame below should load, and 'ALLOWALL' should be accepted as a valid header.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: This text should show up.
